Removed watercolour of geese due to copyright issues
The creator of the Indian Goose (Elephant Room Studio) answered my polite
question whether it is okay to use their creations in this game with
an angry letter, insulting me and accusing me of theft.
This commit removes the birds.
Please don't use any version of the game before this commit.