Added dolphin, changed license terms for set-sizes-animals.html
A dolphin was added as a second animal.

Rachel Nabors, the creator of Tuna (the tomcat) has granted the author of
set-sizes-animals.html, Jacob Kanev (jkanev), the permission to use the
drawing, provided the attribution to Rachel Nabors is never removed,
the software is always used in a non-commercial setting,
and Tuna is removed when Tuna's author requests this.
set-sizes-animals.html is licensed under CC-BY NC ND 4.0 license.
No versions without the copyright statement or without the attribution of
Tuna to Rachel Nabors may be used or distributed.
Should the license terms or subjects of any two versions contradict,
any newer version submitted by Jacob Kanev (jkanev) will override and
invalidate the license terms or subjects of earlier versions.