Switched arguments of mod and log around
The arguments of mod and log are switched round. Mod(a,b) now means
subtract the nearest multiple of b from a, and Log(a,b) means base b
logarithm of a. (It was the other way round before, and I found this
much harder in playing). Also, the debug strings in the hints were
removed.